requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
opencv-python-headless==4.7.0.72
numpy==1.23.5
onnxruntime==1.15.0
rembg==2.0.28
Pillow==9.5.0
